+++ src/pages/Portal/Portal.js
@@ -11,7 +11,8 @@ import WidgetGrid from '../../organisms/WidgetGrid/WidgetGrid';
 
 import Widget from '../../molecules/Widget/Widget';
 
-import BoltIcon from '../../atoms/BoltIcon/BoltIcon';
+import BoltArrowUpIcon from '../../atoms/BoltArrowUpIcon/BoltArrowUpIcon';
+import BoltArrowDownIcon from '../../atoms/BoltArrowDownIcon/BoltArrowDownIcon';
 import WaterIcon from '../../atoms/WaterIcon/WaterIcon';
 import FireIcon from '../../atoms/FireIcon/FireIcon';
 
@@ -33,16 +34,16 @@ export default function Portal() {
 				<Link to="/login"><Button type="primary">Login</Button></Link>
 			</div>
 			<WidgetGrid>
-				<Widget title="Current" name="current-electricity-usage" icon={<BoltIcon />}>
+				<Widget title="Receiving" name="current-electricity-usage" icon={<BoltArrowDownIcon />}>
 					<CurrentElectricityReceived />
 				</Widget>
-				<Widget title="Current" name="current-electricity-delivered" icon={<BoltIcon />}>
+				<Widget title="Delivering" name="current-electricity-delivered" icon={<BoltArrowUpIcon />}>
 					<CurrentElectricityDelivered />
 				</Widget>
-				<Widget title="Today" name="electricity-usage" icon={<BoltIcon />}>
+				<Widget title="Today" name="electricity-usage" icon={<BoltArrowDownIcon />}>
 					<ElectricityReceived />
 				</Widget>
-				<Widget title="Today" name="electricity-delivered" icon={<BoltIcon />}>
+				<Widget title="Today" name="electricity-delivered" icon={<BoltArrowUpIcon />}>
 					<ElectricityDelivered />
 				</Widget>
 				<Widget title="Today" name="gas-usage" icon={<FireIcon />}>
@@ -51,7 +52,7 @@ export default function Portal() {
 				<Widget title="Today" name="water-usage" icon={<WaterIcon />}>
 					<WaterUsage />
 				</Widget>
-				<Widget title="Month" name="electrical-usage-by-month" icon={<BoltIcon />}>
+				<Widget title="Month" name="electrical-usage-by-month" icon={<BoltArrowDownIcon />}>
 					<ElectricityUsageMonth />
 				</Widget>
 				<Widget title="Month" name="gas-usage-by-month" icon={<FireIcon />}>