+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Updated dependencies.
+- Updated existing portal page widgets to be the same as the dashboard page.
 
 ## [1.5.0] - 2021-04-18
 ### Added